follow Rehman's updates and restructure the entire application.
