+++ adviseme/templates/about.html
@@ -1,0 +1,91 @@
+{% extends "layout.html"%}
+
+{% block content %}
+<!--  About Us Page -->
+	<header class="bg-steel text-center py-5 mb-3 center" >
+		<div class="container">
+			<h1 class="font-weight-light text-white">Meet the Team</h1>
+		</div>
+	</header>
+
+	<div class="row content-section">
+		<div class="card mb-3" style="max-width: 1000px;">
+			<div class="row no-gutters">
+
+				<div class="col-md-4">
+					<img src={{ url_for('static', filename="dev_pics/CM.png") }} width="150" class="rounded-left" alt="...">
+				</div>
+				<div class="col-md-8">
+					<div class="card-body">
+						<h5 class="card-title">Chrystal Mingo</h5>
+						<p class="card-text">
+							is a senior majoring in Computer Science at The City College of New York. She has interned at various companies such as Verizon, Morgan Stanley, and Citi as a software developer. She also has a passion for teaching and has worked as a Teacher for the Girls Who Code Summer Immersion Program, and taught Algebra I and II, as well as Intro and AP Computer Science courses in Spanish and English at Gregorio Luperon High School. Chrystal Mingo is also a Grace Hopper 2019 Scholarship Recipient, and current President of Women in Computer. She will be graduating from CCNY in Spring 2021 and entering Citi’s EIOT full-time rotational program as a project manager. She is one of the frontend developers for AdviseMe and believes this project will be a game-changer for advisement at CCNY.
+						</p>
+						<hr>
+						<div style="text-align:center">
+							<a href="https://github.com/chrystalmingo"  target="_blank"> <img src={{ url_for('static', filename="dev_pics/git.png") }} width="30"></a>
+							<a href="#" target="_blank"> <img src={{ url_for('static', filename="dev_pics/linkedin.png") }} width="30" ></a>
+							<a href="#" target="_blank"> <img src={{ url_for('static', filename="dev_pics/gmail.png") }} width="30"></a>
+						</div>
+					</div>
+				</div>
+
+				<div class="col-md-4">
+					<img src={{ url_for('static', filename="dev_pics/ZW.jpg") }} width="150" class="rounded-left" alt="...">
+				</div>
+				<div class="col-md-8">
+					<div class="card-body">
+						<h5 class="card-title">Zhicong Wen</h5>
+						<p class="card-text">
+							is a senior majoring in Computer Science at City College of New York. Previously he was interning as a Python programmer at the NYC Department of Environmental Protection(DEP). He also worked on Quality Assurance of the New York City sewer system database. He will be graduating from CCNY in Fall 2021. He is one of the frontend developers for AdviseMe.
+						</p>
+						<hr>
+						<div style="text-align:center">
+							<a href="https://github.com/zwen000"  target="_blank"> <img src={{ url_for('static', filename="dev_pics/git.png") }} width="30"></a>
+							<a href="https://www.linkedin.com/in/zhicongw-b243a7169/" target="_blank"> <img src={{ url_for('static', filename="dev_pics/linkedin.png") }} width="30" ></a>
+							<a href="#" target="_blank"> <img src={{ url_for('static', filename="dev_pics/gmail.png") }} width="30"></a>
+						</div>
+					</div>
+				</div>
+
+				<div class="col-md-4">
+					<img src={{ url_for('static', filename="dev_pics/XL.jpg") }} width="150" class="rounded-left" alt="...">
+				</div>
+				<div class="col-md-8">
+					<div class="card-body">
+						<h5 class="card-title">Xunshan Lin</h5>
+						<p class="card-text">
+							is a senior majoring in computer engineering at The City College of New York. Previously he was interning as a Python programmer at Mini circuits company. He is beginning to pick up an interest in full-stack development at his senior design project at CCNY. He works as one of the backend developers for AdviseMe and hopes this platform will help students make better academic plans at CCNY.
+						</p>
+						<hr>
+						<div style="text-align:center">
+							<a href="https://github.com/linxunshan" target="_blank"> <img src={{ url_for('static', filename="dev_pics/git.png") }} width="30"></a>
+							<a href="https://www.linkedin.com/in/xunshan-lin-999972205/" target="_blank"> <img src={{ url_for('static', filename="dev_pics/linkedin.png") }} width="30" ></a>
+							<a href="#" target="_blank"> <img src={{ url_for('static', filename="dev_pics/gmail.png") }} width="30"></a>
+						</div>
+					</div>
+				</div>
+
+				<div class="col-md-4">
+					<img src={{ url_for('static', filename="dev_pics/RA.png") }} width="150" class="rounded-left" alt="...">
+				</div>
+				<div class="col-md-8">
+					<div class="card-body">
+						<h5 class="card-title">Rehman Arshad</h5>
+						<p class="card-text">
+							is a senior majoring in Computer Science at The City College of New York. He has interned at various academic research institutions at The Groove school of Engineering, such as NOAA Crest (National Oceanic Atmospheric Administration) and Professor Tarek Sadawi. At NOAA Crest, he worked on data collection and analysis using data from the National Weather Service. With Professor Tarek Sadawi he worked on an IoT medical application that would enable remote patient monitoring as a research assistant. He was a part of CUNY Tech Prep cohort 5, a Full Stack program part of NYC Tech Talent Pipeline. Also, he has a fascination with mathematics and a passion for computer graphics and video game development. He will be graduating from The City College of New York in Fall 2021. He is one of the backend developers for AdviseMe and hopes this online platform can help the next generation of students entering CCNY.
+						</p>
+						<hr>
+						<div style="text-align:center">
+							<a href="https://github.com/rehman000"  target="_blank"> <img src={{ url_for('static', filename="dev_pics/git.png") }} width="30"></a>
+							<a href="https://www.linkedin.com/in/rehman-arshad/" target="_blank"> <img src={{ url_for('static', filename="dev_pics/linkedin.png") }} width="30" ></a>
+							<a href="#" target="_blank"> <img src={{ url_for('static', filename="dev_pics/gmail.png") }} width="30"></a>
+						</div>
+					</div>
+				</div>
+
+			</div>
+		</div>
+	</div>
+
+{% endblock content %}
