+++ adviseme/templates/advisingForm.html
@@ -1,0 +1,6 @@
+{% extends "layout.html" %}
+{% block content %}
+
+<h1>Advising Form Page!</h1>
+
+{% endblock content %}
