+++ adviseme/templates/advisor.html
@@ -1,0 +1,6 @@
+{% extends "layout.html" %}
+{% block content %}
+
+<h1>Advisor Page!</h1>
+
+{% endblock content %}
