+++ adviseme/templates/student.html
@@ -1,0 +1,6 @@
+{% extends "layout.html" %}
+{% block content %}
+
+<h1>Student Page!</h1>
+
+{% endblock content %}
