+++ adviseme/templates/workflow.html
@@ -1,0 +1,6 @@
+{% extends "layout.html" %}
+{% block content %}
+
+<h1>Workflow Page!</h1>
+
+{% endblock content %}
