--- app/templates/about.html
@@ -1,39 +1,0 @@
-{% extends "layout.html"%}
-
-{% block content %}
-<!--    <h1>About Page</h1>-->
-	<header class="bg-steel text-center py-5 mb-3 center" >
-		<div class="container">
-			<h1 class="font-weight-light text-white">Meet the Team</h1>
-		</div>
-	</header>
-
-	 <div class="row content-section">
-
-		 {% for member in team %}
-			 <div class="card mb-3" style="max-width: 1000px;">
-				  <div class="row no-gutters">
-					<div class="col-md-4">
-						<img src={{ member.image }} width="150" class="rounded-left" alt="...">
-					</div>
-
-					<div class="col-md-8">
-					  <div class="card-body">
-						<h5 class="card-title">{{ member.name }}</h5>
-						<p class="card-text">
-							{{ member.description }}
-						</p>
-						  <hr>
-						  <div style="text-align:center">
-							<a href={{ member.git }}  target="_blank"> <img src={{ url_for('static', filename="dev_pics/git.png") }} width="30"></a>
-							<a href={{ member.linkedin }} target="_blank"> <img src={{ url_for('static', filename="dev_pics/linkedin.png") }} width="30" ></a>
-							<a href={{ member.gmail }} target="_blank"> <img src={{ url_for('static', filename="dev_pics/gmail.png") }} width="30"></a>
-						  </div>
-					  </div>
-					</div>
-				  </div>
-				</div>
-		 {% endfor %}
-	</div>
-
-{% endblock content %}
