--- app/templates/advisor.html
@@ -1,6 +1,0 @@
-{% extends "layout.html" %}
-{% block content %}
-
-<h1>Advisor Page!</h1>
-
-{% endblock content %}
