--- app/templates/student.html
@@ -1,6 +1,0 @@
-{% extends "layout.html" %}
-{% block content %}
-
-<h1>Student Page!</h1>
-
-{% endblock content %}
