--- app/templates/workflow.html
@@ -1,6 +1,0 @@
-{% extends "layout.html" %}
-{% block content %}
-
-<h1>Workflow Page!</h1>
-
-{% endblock content %}
