+++ run.py
@@ -1,0 +1,5 @@
+  
+from adviseme import app
+
+if __name__ == '__main__':
+    app.run(debug=True)
